refactor(remove): 将移除功能重构为模块化设计
将原本在app.py中的移除功能代码重构到单独的remove.py模块中，并在task和reward蓝图中添加对应的路由。同时更新前端表单为链接方式，提高用户体验。

重构后的代码结构更清晰，便于维护和扩展。移除功能现在遵循蓝图模式，与添加功能保持一致的代码组织方式。
